Move shader injection to media utils. Attempt to add hover effect to avatar hands
