+++ src/utils/media-highlight-frag.glsl
